Add test data set of ASLR-deficient binaries
As usual the data is made from builds of include-bin with various flag
tweaks. For Windows there are 3 variants: no DYNAMICBASE (daemon),
DYNAMICBASE on but missing relocs (daemonded), and ASLR working, but no
high-entropy VA (daemon-tty). I didn't manage to find a way to reproduce
the missing relocs condition for 32-bit binaries, even though this
happened in the 0.51 release.